don't wait for traces we know aren't coming
